Restricted pages added

+++ base/views.py
@@ -6,6 +6,8 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 
@@ -18,6 +20,9 @@ from django.contrib.auth import authenticate, login, logout
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -74,6 +79,8 @@ def room(request,pk):
 
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -84,10 +91,13 @@ def createRoom(request):
     context={'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form =RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('Your are not allowed!!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -97,9 +107,13 @@ def updateRoom(request,pk):
 
     context={'form': form}
     return render(request, 'base/room_form.html', context)
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host :
+        return HttpResponse('Your are not allowed!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
